% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_data.R
\name{create_test_data}
\alias{create_test_data}
\title{Create test data}
\usage{
create_test_data()
}
\value{
a dummy data.frame of data
}
\description{
Create dataset to be used in tests (useful for demo purposes too)
}
\examples{
df <- create_test_data()
}
