% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ensure_ons_api_available}
\alias{ensure_ons_api_available}
\title{Check if the given ONS json is available}
\usage{
ensure_ons_api_available(ons_api)
}
\arguments{
\item{ons_api}{string; GeoJSON url of a shape file. This can be found on the
ONS geography portal}
}
\value{
\code{TRUE} if the API is available, otherwise an error will occur.
}
\description{
Check if the given ONS json is available
}
