% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fingertipscharts.R
\docType{package}
\name{fingertipscharts}
\alias{fingertipscharts}
\title{fingertipscharts: A package for creating the visualisations displayed in Fingertips}
\description{
The fingertipscharts package provides one type of function
}
\section{quick charts functions}{
 Easy to produce Fingertips charts using
  tidyverse syntax
}

