% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mapdata}
\alias{mapdata}
\title{Dummy data for Upper Tier Local Authorities so dummy maps can be produced}
\format{
A data frame with 152 records and 4 fields:
\describe{
  \item{AreaCode}{Office for National Statistics area codes}
  \item{AreaName}{Office for National Statistics area names}
  \item{Significance}{The statistical significance of the area compared to a benchmark}
  \item{Value}{The indicator value for the data}
}
}
\usage{
mapdata
}
\description{
A dataset of indicator values for all Upper Tier Local Authorities
}
\keyword{datasets}
