% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spine_chart_utils.R
\name{round2}
\alias{round2}
\title{Proper rounding of values}
\usage{
round2(val, dps)
}
\arguments{
\item{val}{numeric value to round}

\item{dps}{numeric, number of decimal places}
}
\description{
Proper rounding of values
}
\details{
function taken from this link (\url{https://stackoverflow.com/questions/12688717/round-up-from-5})
}
