% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{scale_fill_phe}
\alias{scale_fill_phe}
\title{Fingertips scale fill for ggplot2}
\usage{
scale_fill_phe(theme = "fingertips", ...)
}
\arguments{
\item{theme}{string; current options are fingertips only for discrete
scales}

\item{...}{inputs to the scale_manual (for discrete values)}
}
\description{
Fingertips scale fill
}
