% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_phe}
\alias{theme_phe}
\title{Fingertips theme for ggplot2}
\usage{
theme_phe(theme = "fingertips", base_size = 11, base_family = "")
}
\arguments{
\item{theme}{string; theme of chart, current are fingertips only}

\item{base_size}{base font size}

\item{base_family}{base font family}
}
\description{
fingertips theme
}
