% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ensure_ons_api_available}
\alias{ensure_ons_api_available}
\title{Check if the given ONS json is available}
\usage{
ensure_ons_api_available(ons_api)
}
\arguments{
\item{endpoint}{string, the json to check}
}
\value{
\code{TRUE} if the API is available, otherwise \code{stop()} is called.
}
\description{
Check if the given ONS json is available
}
